# Optimization for Float.round/2 when precision is 0