I did not notice one more optimization opportunity in previous PR. 

Benchmarks (1000 floats per run)

Before
```
Name                          ips        average  deviation         median         99th %
round/1                    382.08     2617.25 μs     ±3.97%        2590 μs     3031.20 μs
round/2 precision 0        377.92     2646.05 μs     ±4.09%        2621 μs     3064.80 μs
round/2 precision 1        351.24     2847.02 μs     ±6.13%        2804 μs     3648.76 μs
```

After
```
Name                          ips        average  deviation         median         99th %
round/1                  20596.03       48.55 μs    ±18.17%          47 μs          91 μs
round/2 precision 0      20096.23       49.76 μs    ±17.50%          48 μs          91 μs
round/2 precision 1        354.77     2818.69 μs     ±4.28%        2795 μs     3292.34 μs
```